fix some foo from the ext_header changes
notably, while im sure 0xffff is _supposed_ to be -1, we are converting to a wider type
we have checks/assignments everywhere specifically for 0xffff so its not feasible to change quickly

+++ gemrb/core/Item.h
@@ -261,7 +261,7 @@ public:
 	{
 		if(which < 0)
 			return GetWeaponHeader(which == -2) ;
-		if (ext_headers.size() <= size_t(which)) {
+		if (int(ext_headers.size()) <= which) {
 			return NULL;
 		}
 		return &ext_headers[which];
